#40 Converted Comment components into functional component
