--- frontend/src/components/Comment/Comment.scss
@@ -1,10 +1,0 @@
-.comment-div {
-  border-bottom: 1px solid grey;
-  padding-left: 20px;
-  font-size: 14px;
-}
-
-.date-created {
-  color: grey;
-  font-size: 12px;
-}
