--- frontend/src/components/Comment/PostComment.scss
@@ -1,8 +1,0 @@
-.comment-form-div {
-  display: flex;
-  height: 30px;
-}
-
-.comment-input {
-  flex-grow: 1;
-}
